requirements.txt
phidata
pydantic
python-dotenv
duckduckgo-search
sqlalchemy
fastapi
uvicorn
newspaper3k
arxiv
pypdf
duckdb
googlesearch-python
pycountry
lxml-html-clean
lancedb
tantivy
ollama
phi-agent
phi-model-ollama

